requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
instagrapi==1.19.4
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pydantic==1.10.9
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
Pillow>=8.1.1
requests>=2.28.0
